removed get_data_by_name from extract to connector
