+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - using `OrdicPerson`'s orcidid.uri for transformation instead stitiching it together
 - fetch method in OrcidConnector accepts directly filters + passes query als param to search-endpoint
 - moved `get_data_by_id` from extract to connector
+- move `get_data_by_name` from extract to connector
 ### Deprecated
 
 ### Removed
